2 Refactor `QuiteOkImage.at()` to return boxed protocol type
I am not overly pleased with the term `Pixel` for the protocol, since
it has no information on the used color model, but I am unable to think
of a better name for that protocol.

I am also unsure if I should be further defining the Protocol; at this
moment, it really just is a wrapper over both colour models.