1 Wrap `phoboslab/qoi/qoi_decode()`
At this point, the wraper is as minimal as possible, without
consideration to performance.

`QOI_OP_RGB` and `QOI_OP_RGBA` would be sufficiently tested with a single
pixel, since each pixel is independent of previous ones; testing also
multiple pixels serves as control to ensure an image with more than one
single pixel is properly decoded; this control is not made for all other
operations.

`decodeImageWithIndexToFullArray()` is not implemented, because I think
it is not necessary to validate the decoder, since we already check both
that a (partially) filled array, and also that an overwritten value are
both properly decoded.

`QOI_OP_DIFF` and `QOI_OP_LUMA` test again a single pixel, because these
always refer to the previous pixel only; it is not important for these
operations to know how many other pixels might exist in the image.

Some of these tests could be parametrised; I did not dive deep into how
to create such tests, and it is not necessary to parametrise them now.